Invocie-Auditor/requirements.txt
langgraph
langchain
langchain-community
litellm
fastapi
uvicorn
pydantic
pypdf
pdfplumber
python-docx
pytesseract
Pillow
faiss-cpu
sentence-transformers
pyyaml
watchdog
streamlit
boto3
requests
